Move process tree types to common folder
