+++ x-pack/plugins/session_view/public/components/ProcessTreeNode/index.tsx
@@ -14,7 +14,7 @@
 import React, { useMemo, useRef, useLayoutEffect, useState, useEffect, MouseEvent } from 'react';
 import { EuiButton, EuiIcon } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n/react';
-import { Process } from '../../hooks/use_process_tree';
+import { Process } from '../../../common/types/process_tree';
 import { useStyles, ButtonType } from './styles';
 import { ProcessTreeAlerts } from '../ProcessTreeAlerts';
