+++ x-pack/plugins/session_view/public/components/SessionView/hooks.ts
@@ -9,7 +9,7 @@ import { useQuery } from 'react-query';
 import { EuiSearchBarOnChangeArgs } from '@elastic/eui';
 import { CoreStart } from 'kibana/public';
 import { useKibana } from '../../../../../../src/plugins/kibana_react/public';
-import { ProcessEvent } from '../../hooks/use_process_tree';
+import { ProcessEvent } from '../../../common/types/process_tree';
 import { PROCESS_EVENTS_ROUTE } from '../../../common/constants';
 
 interface ProcessEventResults {
